+++ README.md
@@ -2,28 +2,34 @@
 
 The `DafnyVMC` module introduces utils for probabilistic reasoning in Dafny. At the moment, the API is intentionally limited in scope, and only supports compilation to Java. For the future, we plan to extend both the functionality and the range of supported languages.
 
-There are two ways to use the library:
-- The trait `DRandomFoundational` only needs fair coin flips as a primitive.
-- The trait `DRandomExternUniformPowerOfTwo` also uses an external implementation of uniformly distributed random numbers.
+# Dafny Examples
 
-To run the examples in the `docs` directory, use one of the following commands:
+To run the examples in the `docs/dafny` directory, use the following commands:
 
 ```bash
-# Java Examples Foundational
-$ dafny build docs/dafny/ExamplesFoundational.dfy --target:java src/interop/java/DRandomCoin.java src/interop/java/DRandomUniformPowerOfTwo.java dfyconfig.toml --no-verify
-$ java -jar docs/dafny/ExamplesFoundational.jar
-# Java Examples ExternUniformPowerOfTwo
-$ dafny build docs/dafny/ExamplesExternUniformPowerOfTwo.dfy --target:java src/interop/java/DRandomCoin.java src/interop/java/DRandomUniformPowerOfTwo.java dfyconfig.toml --no-verify
-$ java -jar docs/dafny/ExamplesExternUniformPowerOfTwo.jar
+# Dafny Examples
+$ dafny build docs/dafny/ExamplesRandom.dfy --target:java src/interop/java/Random.java dfyconfig.toml --no-verify
+$ java -jar docs/dafny/ExamplesRandom.jar
+```
+
+# Java Examples
+
+To run the examples in the `docs/java` directory, use the following command:
+
+```bash
+# Java Examples
+$ bash scripts/build.sh 
+$ bash build/java/run.sh  
 ```
 
 # Testing
 
-To run the statistical tests in the `tests` diretory, use one of the following commands:
+To run the statistical tests in the `tests` directory, use the following commands:
 
 ```bash
-# Java Tests Foundational
-$ dafny test --target:java src/interop/java/DRandomCoin.java src/interop/java/DRandomUniformPowerOfTwo.java tests/TestsFoundational.dfy tests/Tests.dfy dfyconfig.toml --no-verify
-# Java Tests ExternUniformPowerOfTwo
-$ dafny test --target:java src/interop/java/DRandomCoin.java src/interop/java/DRandomUniformPowerOfTwo.java tests/TestsExternUniformPowerOfTwo.dfy tests/Tests.dfy dfyconfig.toml --no-verify
+# Dafny Tests
+$ dafny test --target:java src/interop/java/Random.java tests/TestsRandom.dfy tests/Tests.dfy dfyconfig.toml --no-verify
 ```
+
+
+
