+++ scripts/test.sh
@@ -15,23 +15,13 @@ then
 fi
 
 echo Running $TARGET_LANG tests...
-echo "Running tests/TestsRandomCoin.dfy:"
-time $DAFNY test --target:$TARGET_LANG src/interop/$TARGET_LANG/Minus/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/Minus/Random.$TARGET_LANG src/interop/$TARGET_LANG/Random.$TARGET_LANG tests/TestsRandomCoin.dfy tests/Tests.dfy dfyconfig.toml --no-verify
 echo "Running tests/TestsRandom.dfy:"
-time $DAFNY test --target:$TARGET_LANG src/interop/$TARGET_LANG/Minus/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/Minus/Random.$TARGET_LANG src/interop/$TARGET_LANG/Random.$TARGET_LANG tests/TestsRandom.dfy tests/Tests.dfy dfyconfig.toml  --no-verify
+time $DAFNY test --target:$TARGET_LANG src/interop/$TARGET_LANG/Random.$TARGET_LANG tests/TestsRandom.dfy tests/Tests.dfy dfyconfig.toml  --no-verify
 
 echo Running $TARGET_LANG documentation...
 
-echo "Building docs/dafny/ExamplesRandomCoin.dfy..." 
-$DAFNY build docs/dafny/ExamplesRandomCoin.dfy --target:$TARGET_LANG src/interop/$TARGET_LANG/Minus/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/Minus/Random.$TARGET_LANG src/interop/$TARGET_LANG/Random.$TARGET_LANG  dfyconfig.toml --no-verify
-echo "Executing compiled docs/dafny/ExamplesRandomCoin.dfy:" 
-if [ "$TARGET_LANG" = "java" ]
-then
-  java -jar docs/dafny/ExamplesRandomCoin.jar
-fi
-
 echo "Building docs/dafny/ExamplesRandom.dfy..." 
-$DAFNY build docs/dafny/ExamplesRandom.dfy --target:$TARGET_LANG src/interop/$TARGET_LANG/Minus/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/RandomCoin.$TARGET_LANG src/interop/$TARGET_LANG/Minus/Random.$TARGET_LANG src/interop/$TARGET_LANG/Random.$TARGET_LANG  dfyconfig.toml --no-verify
+$DAFNY build docs/dafny/ExamplesRandom.dfy --target:$TARGET_LANG src/interop/$TARGET_LANG/Random.$TARGET_LANG dfyconfig.toml --no-verify
 echo "Executing compiled docs/dafny/ExamplesRandom.dfy:" 
 if [ "$TARGET_LANG" = "java" ]
 then
